modify: kakaoMap 훅으로 분리
